Moved betaerror stuff into C/betaerror.c. Added debug info to beta.dumps when linked with betarun_debug
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@7375 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/valhallaComm.c
@@ -829,7 +829,7 @@ static int valhallaCommunicate (int PC, int SP, Object* curObj)
     case VOP_MAIN_PHYSICAL: {
       DEBUG_VALHALLA(fprintf(output,
 			     "VOP_MAIN_PHYSICAL()=0x%x",
-			     getMainPhysical()));
+			     (int)getMainPhysical()));
       valhalla_writeint(opcode);
       valhalla_writeint(getMainPhysical());
       valhalla_socket_flush ();
